Installed Toastr Notifications and added error indicators when attempting to create a new record in Airtable

+++ src/services/movieService.js
@@ -2,7 +2,7 @@
 const BASE_URL = `http://www.omdbapi.com/?apikey=c7933753`;
 const AIRTABLE_BASE = `appiKbPo7d5P9pnCG`;
 const AIRTABLE_TABLE = `tblcMlwFSdnAcvIvt`;
-const AIRTABLE_API_KEY = `pathY8nl4S5cr40VQ.585f4ba1783e408435d65b4367715394bcd6ae2243f4cc80a5ece02739996070`;
+const AIRTABLE_API_KEY = `pathY8nl4S5cr40VQ.a7a82c062d2589a19174ef0ca73199fea3e5d2cfc3d08ba24b06d09792d57d23`;
 const AIRTABLE_URL = `https://api.airtable.com/v0/${AIRTABLE_BASE}/${AIRTABLE_TABLE}`;
 
 // Authorization Data for Airtable
@@ -67,7 +67,7 @@ async function create(movie) {
   try {
     const response = await fetch(AIRTABLE_URL, {
       method: "POST",
-      headersData,
+      headers: headersData,
       body: JSON.stringify({
         fields: {
           Title: movie.Title,
@@ -94,20 +94,37 @@ async function create(movie) {
   }
 }
 
-// Delete Favourite Movies Data from Airtable
+// Delete Favourite Movies Data from Airtable by using imdbID to fetch the record ID
+// Airtable can only delete records via the record ID
 async function destroy(imdbID) {
   try {
-    const response = await fetch(AIRTABLE_URL + `/${imdbID}`, {
+    const response = await fetch(
+      `${AIRTABLE_URL}?filterByFormula={imdbID}="${imdbID}"`,
+      {
+        headers: headersData,
+      }
+    );
+    if (!response.ok) {
+      throw new Error(`Response status: ${response.status}`);
+    }
+    const json = await response.json();
+    const recordId = fetchData.records[0]?.id; // Gets the record ID
+    if (!recordId) {
+      throw new Error("Record not found");
+    }
+    const deleteRecord = await fetch(`${AIRTABLE_URL}/${recordId}`, {
       method: "DELETE",
       headers: {
         Authorization: `Bearer ${AIRTABLE_API_KEY}`,
       },
     });
-    if (!response.ok) {
-      throw new Error(`Response status: ${response.status}`);
+
+    if (!deleteRecord.ok) {
+      throw new Error(`Delete Response status: ${deleteRecord.status}`);
     }
-    const json = await response.json();
-    return json;
+
+    const deleteData = await deleteRecord.json();
+    return deleteData;
   } catch (error) {
     console.error(error.message);
   }
